refactor: 패키지 구조 수정 - session/enrollment로 패키지 생성
