+++ src/main/java/nextstep/courses/domain/enrollment/FreeEnrollmentPolicy.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.enrollment;
 
 import nextstep.payments.domain.Payment;
