+++ src/main/java/nextstep/courses/domain/enrollment/PolicyType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.enrollment;
 
 public enum PolicyType {
     FREE,
