+++ src/main/java/nextstep/courses/domain/session/cover/CoverImage.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session.cover;
 
 public class CoverImage {
     private final ImageSize imageSize;
