+++ src/main/java/nextstep/courses/domain/session/cover/ImageType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session.cover;
 
 import java.util.Arrays;
