+++ src/main/java/nextstep/courses/infrastructure/JdbcEnrollmentRepository.java
@@ -1,6 +1,6 @@
 package nextstep.courses.infrastructure;
 
-import nextstep.courses.domain.Enrollment;
+import nextstep.courses.domain.enrollment.Enrollment;
 import nextstep.courses.repository.EnrollmentRepository;
 import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.stereotype.Repository;
