+++ src/main/java/nextstep/courses/repository/SessionRepository.java
@@ -1,6 +1,6 @@
 package nextstep.courses.repository;
 
-import nextstep.courses.domain.Session;
+import nextstep.courses.domain.session.Session;
 
 public interface SessionRepository {
     int save(Session session);
