+++ src/test/java/nextstep/courses/domain/CapacityTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.enrollment.Capacity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
