+++ src/test/java/nextstep/courses/domain/EnrollmentsTest.java
@@ -1,5 +1,7 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.enrollment.Capacity;
+import nextstep.courses.domain.enrollment.Enrollments;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
