+++ src/test/java/nextstep/courses/domain/ImageDimensionTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.session.cover.ImageDimension;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
