+++ src/test/java/nextstep/courses/domain/SessionDurationTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.session.SessionDuration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
