+++ src/test/java/nextstep/courses/domain/SessionTest.java
@@ -1,5 +1,7 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.enrollment.Money;
+import nextstep.courses.domain.session.Session;
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
