+++ src/test/java/nextstep/courses/infrastructure/EnrollmentRepositoryTest.java
@@ -1,6 +1,6 @@
 package nextstep.courses.infrastructure;
 
-import nextstep.courses.domain.Enrollment;
+import nextstep.courses.domain.enrollment.Enrollment;
 import nextstep.courses.repository.EnrollmentRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
